# app/src/app.ts
import { Hono } from 'hono';
import { AppError, PreconditionFailedError, ValidationError } from './lib/errors';
import { createHealthRoutes } from './routes/health';
import { createMapsRoutes, type MapsRouteDependencies } from './routes/maps';
import { createSchemaRoutes } from './routes/schema';

export type AppDependencies = MapsRouteDependencies;

type AppErrorStatus = 400 | 404 | 409 | 412;

export function createApp(dependencies: AppDependencies): Hono {
  const app = new Hono();

  app.onError((error, c) => {
    if (error instanceof AppError) {
      if (error instanceof ValidationError) {
        return c.json(
          {
            error: 'VALIDATION_ERROR',
            message: error.message,
            details: error.details,
          },
          error.statusCode as AppErrorStatus
        );
      }

      return c.json(
        {
          error: error.code,
          message: error.message,
          ...(error instanceof PreconditionFailedError && error.currentVersion !== undefined
            ? { currentVersion: error.currentVersion }
            : {}),
        },
        error.statusCode as AppErrorStatus
      );
    }

    console.error(error);
    return c.json(
      {
        error: 'INTERNAL_SERVER_ERROR',
        message: 'Unexpected server error',
      },
      500
    );
  });

  app.notFound((c) => {
    return c.json(
      {
        error: 'NOT_FOUND',
        message: 'Resource not found',
      },
      404
    );
  });

  app.get('/', (c) => c.redirect('/schema/swagger-ui'));
  app.route('/', createHealthRoutes());
  app.route('/', createSchemaRoutes());
  app.route('/api/v1/maps', createMapsRoutes(dependencies));

  return app;
}